# src/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Babeonym API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "schemas": {
...
      "GivenName": {
        "type": "object",
        "required": [
          "givenName",
          "givenCustomNameBridgeId",
          "rating"
        ],
        "properties": {
          "givenName": {
            "type": "string",
            "example": "Oliver"
          },
          "givenCustomNameBridgeId": {
            "type": "number",
            "example": 42
          },
          "rating": {
            "type": "number",
            "example": 1000
          },
          "percentile": {
            "type": "number",
            "nullable": true,
            "example": 92
          },
          "gender": {
            "type": "string",
            "nullable": true,
            "description": "Key of Gender enum",
            "example": "male"
          }
        }
      },
      "NotAuthenticatedResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Not authenticated"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/givenName/candidates": {
      "get": {
        "operationId": "v1GivenNameCandidates",
        "summary": "Get given name candidates",
        "description": "Returns candidate given names for the authenticated user based on optional filters.",
        "tags": [
          "GivenName"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "popularity",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Popularity filter"
          },
          {
            "in": "query",
            "name": "genders",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated list of genders"
          },
          {
            "in": "query",
            "name": "decadeIds",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated decade IDs"
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Maximum number of results"
          },
          {
            "in": "query",
            "name": "include",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Comma-separated include options (e.g. meta)"
          }
        ],
        "responses": {
          "200": {
            "description": "Candidate given names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GivenName"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/givenName/compare": {
      "post": {
        "operationId": "v1GivenNameCompare",
        "summary": "Compare two given names",
        "description": "Records a comparison result between two given names for the authenticated user.",
        "tags": [
          "GivenName"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "winnerId",
                  "loserId"
                ],
                "properties": {
                  "winnerId": {
                    "type": "number",
                    "example": 12
                  },
                  "loserId": {
                    "type": "number",
                    "example": 34
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comparison recorded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Given name comparison updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid winnerId or loserId",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid winnerId or loserId"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/givenName/custom": {
      "post": {
        "operationId": "v1GivenNameCustom",
        "summary": "Add a custom given name",
        "description": "Adds a user-defined custom given name for the authenticated user.",
        "tags": [
          "GivenName"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "customGivenName"
                ],
                "properties": {
                  "customGivenName": {
                    "type": "string",
                    "example": "Aurelius"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Custom given name added successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Custom given name added successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or inappropriate custom given name",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/givenName/action": {
      "post": {
        "operationId": "v1GivenNameAction",
        "summary": "Update given name state",
        "description": "Updates the state of a given name (approved, rejected, or snoozed) for the authenticated user.",
        "tags": [
          "GivenName"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "givenCustomNameBridgeId",
                  "newState"
                ],
                "properties": {
                  "givenCustomNameBridgeId": {
                    "type": "number",
                    "example": 123
                  },
                  "newState": {
                    "type": "string",
                    "enum": [
                      "APPROVED",
                      "REJECTED",
                      "SNOOZED"
                    ],
                    "example": "APPROVED"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Given name action updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Given name action updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Invalid request parameters"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotAuthenticatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}